ultimos fixes
inputs flotantes, engordes predefs

+++ src/actions/action-engorde.js
@@ -73,7 +73,7 @@ export function InputVariacionPastureValor(id, pagina ,valor) {
     return {
         type: "VALORVARIACION_PASTURE_ENGORDE",
         posicion:id,
-        valor:parseInt(valor)
+        valor:parseFloat(valor)
     }    
 }
 
@@ -82,7 +82,7 @@ export function InputVariacionGrainValor(id, pagina, valor) {
     return {
         type: "VALORVARIACION_GRAIN_ENGORDE",
         posicion:id,
-        valor:parseInt(valor)
+        valor:parseFloat(valor)
     }    
 }
 
@@ -90,7 +90,7 @@ export function InputVariacionSilageValor(id, pagina, valor) {
     return {
         type: "VALORVARIACION_SILAGE_ENGORDE",
         posicion:id,
-        valor:parseInt(valor)
+        valor:parseFloat(valor)
     }    
 }
 
@@ -98,7 +98,7 @@ export function InputVariacionRastrojoValor(id, pagina, valor) {
     return {
         type: "VALORVARIACION_RASTROJO_ENGORDE",
         posicion:id,
-        valor:parseInt(valor)
+        valor:parseFloat(valor)
     }    
 }
 
@@ -106,7 +106,7 @@ export function InputVariacionDiferidosValor(id, pagina, valor) {
     return {
         type: "VALORVARIACION_DIFERIDOS_ENGORDE",
         posicion:id,
-        valor:parseInt(valor)
+        valor:parseFloat(valor)
     }    
 }
